+++ .github/workflows/gradle-build.yml
@@ -9,7 +9,7 @@ name: Gradle Build
 
 on:
   push:
-    branches: [ "user", "schedule", "deploy" ]
+    branches: [ "user", "schedule", "deploy", "utility" ]
 
 permissions:
   contents: read
