+++ src/main/java/org/standard/dreamcalendar/domain/user/UserService.java
@@ -6,14 +6,14 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.standard.dreamcalendar.domain.user.type.Role;
-import org.standard.dreamcalendar.service.Encryptor;
-import org.standard.dreamcalendar.service.JwtProvider;
+import org.standard.dreamcalendar.util.Encryptor;
+import org.standard.dreamcalendar.util.JwtProvider;
 import org.standard.dreamcalendar.domain.user.dto.response.UpdateTokenResponse;
 import org.standard.dreamcalendar.domain.user.type.TokenType;
 import org.standard.dreamcalendar.domain.user.dto.response.LogInByEmailPasswordResponse;
 import org.standard.dreamcalendar.domain.user.dto.UserDto;
 import org.standard.dreamcalendar.domain.user.type.TokenValidationStatus;
-import org.standard.dreamcalendar.service.DtoConverter;
+import org.standard.dreamcalendar.util.DtoConverter;
 
 import java.security.NoSuchAlgorithmException;
 import java.util.List;
